refactor(dashboard): split oversized IDE editor components into focused modules
Extract from ide-editor-extensions.ts (1429 → ~290 lines):
- ide-editor-theme.ts         - CM6 theme + syntax highlighting
- ide-editor-language.ts      - OCaml/TOML/YAML StreamLanguage + dynamic loaders
- ide-editor-trace-gutter.ts  - Keeper trace gutter + chip extensions
- ide-editor-ownership.ts     - Blame gutter + ownership StateEffect

Extract from ide-editor.ts (923 lines removed):
- ide-editor-annotation-ui.ts - AnnotationPopover, context route links
- ide-editor-blame.ts         - BlameTimeline, layer overlay summary
- ide-editor-cursor.ts        - Keeper cursor chips
- ide-editor-find.ts          - Find/replace panel
- ide-editor-signals.ts       - File signals, trace line focus

Both original files remain as re-export hubs for backward compatibility.
