+++ dashboard/src/components/ide/ide-editor-extensions.ts
@@ -1,16 +1,6 @@
-import { Decoration, EditorView, GutterMarker, WidgetType, gutter, lineNumbers, type DecorationSet, type ViewUpdate } from '@codemirror/view'
+import { Decoration, EditorView, WidgetType, lineNumbers, type DecorationSet } from '@codemirror/view'
 import { Annotation, EditorState, Extension, RangeSetBuilder, StateField, StateEffect, type Text } from '@codemirror/state'
-import {
-  defaultHighlightStyle,
-  StreamLanguage,
-  syntaxHighlighting,
-  type StringStream,
-  type StreamParser,
-} from '@codemirror/language'
 import type { LineOwnership } from './keeper-line-ownership-store'
-import type { KeeperTraceContextFields, KeeperTraceEvent, KeeperTraceSource } from './keeper-trace-store'
-import { kSigil } from '../keeper-badge'
-import { isPositiveSafeInteger } from '../common/normalize'
 
 // ── Read-only lock ────────────────────────────────────────────────
 // Prevents all user input. CM6 6.x uses EditorState.changeFilter.
@@ -23,786 +13,31 @@ export function readOnlyExt(): Extension {
   )
 }
 
-// ── Theme from design-system CSS variables ────────────────────────
-// Maps semantic tokens to CM6 theme facets so the editor matches the
-// dashboard light/dark theme without a hardcoded theme object.
+// ── Re-exports from extracted modules ─────────────────────────────
+// Theme, language, trace gutter, and ownership were extracted to
+// keep this file under the 300-line threshold.
 
-export function themeExt(): Extension {
-  return EditorView.theme({
-    '&': {
-      background: 'var(--color-bg-page)',
-      color: 'var(--color-fg-secondary)',
-      fontFamily: 'var(--font-mono)',
-      fontSize: 'var(--fs-13)',
-      lineHeight: '1.6',
-      height: '100%',
-    },
-    '.cm-scroller': {
-      overflow: 'auto',
-      minHeight: '0',
-    },
-    '.cm-content': {
-      caretColor: 'transparent',
-      padding: 'var(--sp-2) 0',
-      minHeight: '100%',
-    },
-    '.cm-line': {
-      padding: '0 var(--sp-3)',
-    },
-    '.cm-gutters': {
-      background: 'var(--color-bg-page)',
-      color: 'var(--color-fg-disabled)',
-      border: 'none',
-      fontSize: 'var(--fs-11)',
-      minWidth: '44px',
-      position: 'sticky',
-      left: '0',
-      zIndex: '2',
-    },
-    '.cm-gutterElement': {
-      textAlign: 'right',
-      paddingRight: 'var(--sp-2)',
-    },
-    '.cm-lineNumbers': {
-      borderRight: '1px solid var(--color-border-default)',
-    },
-    '.cm-blame-gutter': {
-      minWidth: '78px',
-      borderRight: '1px solid var(--color-border-default)',
-      background: 'var(--color-bg-surface)',
-    },
-    '.cm-blame-gutter .cm-gutterElement': {
-      padding: '0 var(--sp-2)',
-      textAlign: 'left',
-    },
-    '.cm-blame-marker': {
-      display: 'inline-grid',
-      gridTemplateColumns: '18px minmax(0, 1fr)',
-      alignItems: 'center',
-      gap: 'var(--sp-1)',
-      width: '68px',
-      minWidth: '0',
-      color: 'var(--cm-blame-color)',
-      fontSize: 'var(--fs-10)',
-      lineHeight: '1.4',
-    },
-    '.cm-blame-sigil': {
-      display: 'inline-flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      width: '16px',
-      height: '14px',
-      borderRadius: 'var(--r-0)',
-      color: 'var(--color-bg-page)',
-      background: 'var(--cm-blame-color)',
-      fontSize: 'var(--fs-9)',
-      fontWeight: '700',
-      letterSpacing: '0',
-    },
-    '.cm-blame-name': {
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      whiteSpace: 'nowrap',
-    },
-    '.cm-trace-gutter': {
-      minWidth: '34px',
-      borderRight: '1px solid var(--color-border-default)',
-      background: 'var(--color-bg-page)',
-    },
-    '.cm-trace-gutter .cm-gutterElement': {
-      padding: '0 var(--sp-1)',
-      textAlign: 'left',
-    },
-    '.cm-trace-stack': {
-      display: 'inline-flex',
-      alignItems: 'center',
-      justifyContent: 'flex-start',
-      gap: '2px',
-      minWidth: '28px',
-      height: '100%',
-      padding: '0',
-      border: '0',
-      background: 'transparent',
-      color: 'inherit',
-      cursor: 'pointer',
-    },
-    '.cm-trace-stack:focus-visible': {
-      outline: '1px solid var(--color-accent-fg)',
-      outlineOffset: '1px',
-    },
-    '.cm-trace-dot': {
-      display: 'inline-block',
-      width: '6px',
-      height: '6px',
-      borderRadius: '999px',
-      background: 'var(--cm-trace-color)',
-      boxShadow: '0 0 0 1px var(--color-bg-page)',
-    },
-    '.cm-trace-overflow': {
-      color: 'var(--color-fg-muted)',
-      fontSize: 'var(--fs-9)',
-      lineHeight: '1',
-    },
-    '&.cm-focused': {
-      outline: 'none',
-    },
-    '.cm-cursor': {
-      display: 'none',
-    },
-    '.cm-selectionBackground': {
-      background: 'transparent !important',
-    },
-    '.cm-line.cm-masc-context-focus': {
-      background: 'color-mix(in srgb, var(--color-accent-fg) 12%, transparent)',
-      boxShadow: 'inset 2px 0 0 var(--color-accent-fg)',
-    },
-    '.cm-masc-context-focus-chip': {
-      display: 'inline-flex',
-      alignItems: 'center',
-      maxWidth: '36ch',
-      marginLeft: 'var(--sp-2)',
-      padding: '0 var(--sp-2)',
-      border: '1px solid var(--color-border-default)',
-      borderRadius: 'var(--r-0)',
-      background: 'var(--color-bg-elevated)',
-      color: 'var(--color-accent-fg)',
-      fontFamily: 'var(--font-mono)',
-      fontSize: 'var(--fs-10)',
-      lineHeight: '1.4',
-      verticalAlign: 'baseline',
-      whiteSpace: 'nowrap',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      pointerEvents: 'none',
-      userSelect: 'none',
-    },
-    '.cm-masc-annotation-chip': {
-      display: 'inline-flex',
-      alignItems: 'center',
-      maxWidth: '34ch',
-      marginLeft: 'var(--sp-2)',
-      padding: '0 var(--sp-2)',
-      border: '1px solid var(--color-border-default)',
-      borderRadius: 'var(--r-0)',
-      background: 'var(--color-bg-surface)',
-      color: 'var(--color-fg-muted)',
-      fontFamily: 'var(--font-mono)',
-      fontSize: 'var(--fs-10)',
-      lineHeight: '1.4',
-      verticalAlign: 'baseline',
-      whiteSpace: 'nowrap',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      pointerEvents: 'none',
-      userSelect: 'none',
-    },
-    '.cm-masc-trace-chip': {
-      display: 'inline-flex',
-      alignItems: 'center',
-      maxWidth: '46ch',
-      marginLeft: 'var(--sp-2)',
-      padding: '0 var(--sp-2)',
-      border: '1px solid var(--color-border-default)',
-      borderRadius: 'var(--r-0)',
-      background: 'var(--color-bg-elevated)',
-      boxShadow: 'inset 2px 0 0 var(--cm-trace-chip-color)',
-      color: 'var(--color-fg-secondary)',
-      fontFamily: 'var(--font-mono)',
-      fontSize: 'var(--fs-10)',
-      lineHeight: '1.4',
-      verticalAlign: 'baseline',
-      whiteSpace: 'nowrap',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      pointerEvents: 'none',
-      userSelect: 'none',
-    },
-  })
-}
+export { themeExt, syntaxHighlightExt } from './ide-editor-theme'
+export {
+  languageExt,
+  languageIdForFilePath,
+} from './ide-editor-language'
+export {
+  type EditorKeeperTraceLineEvent,
+  type EditorKeeperTraceLine,
+  type KeeperTraceLineGutterOptions,
+  keeperTraceLineGutterExt,
+  keeperTraceLineChipExt,
+  pushKeeperTraceLines,
+  keeperTraceLinesForFile,
+} from './ide-editor-trace-gutter'
+export {
+  setOwnership,
+  pushOwnership,
+  blameExtensions,
+} from './ide-editor-ownership'
 
-export function syntaxHighlightExt(): Extension {
-  return syntaxHighlighting(defaultHighlightStyle, { fallback: true })
-}
-
-// ── Blame gutter ──────────────────────────────────────────────────
-// Left gutter showing keeper ownership per line. Uses a StateEffect
-// to push ownership updates into the gutter marker.
-
-interface BlameMarkerValue {
-  readonly keeperId: string
-  readonly hueIndex: number
-  readonly editKind: string
-}
-
-class BlameMarker extends GutterMarker {
-  constructor(private readonly info: BlameMarkerValue | null) {
-    super()
-  }
-
-  toDOM(): HTMLElement {
-    const el = document.createElement('span')
-    if (!this.info) {
-      el.textContent = '—'
-      el.style.color = 'var(--color-fg-disabled)'
-      el.style.fontSize = 'var(--fs-11)'
-      return el
-    }
-    const slot = Math.min(12, Math.max(1, this.info.hueIndex || 1))
-    const color = `var(--color-keeper-${slot}, var(--k-${slot}))`
-    el.className = 'cm-blame-marker'
-    el.style.setProperty('--cm-blame-color', color)
-    el.title = `${this.info.keeperId} · ${this.info.editKind}`
-    const sigil = document.createElement('span')
-    sigil.className = 'cm-blame-sigil'
-    sigil.textContent = kSigil(this.info.keeperId)
-    sigil.setAttribute('aria-hidden', 'true')
-    const name = document.createElement('span')
-    name.className = 'cm-blame-name'
-    name.textContent = this.info.keeperId
-    el.append(sigil, name)
-    return el
-  }
-
-  eq(other: BlameMarker): boolean {
-    if (!this.info && !other.info) return true
-    if (!this.info || !other.info) return false
-    return this.info.keeperId === other.info.keeperId && this.info.editKind === other.info.editKind
-  }
-}
-
-const BLAME_EMPTY = new BlameMarker(null)
-
-export const setOwnership = StateEffect.define<ReadonlyMap<number, LineOwnership>>()
-
-const blameMarkerField = StateField.define<BlameMarker[]>({
-  create() {
-    return []
-  },
-  update(markers, tr) {
-    for (const effect of tr.effects) {
-      if (effect.is(setOwnership)) {
-        const ownership = effect.value
-        const newMarkers: BlameMarker[] = []
-        const doc = tr.state.doc
-        for (let i = 1; i <= doc.lines; i++) {
-          const owner = ownership.get(i)
-          if (owner) {
-            newMarkers.push(new BlameMarker({
-              keeperId: owner.keeper_id,
-              hueIndex: owner.hue_index,
-              editKind: owner.last_edit_kind,
-            }))
-          } else {
-            newMarkers.push(BLAME_EMPTY)
-          }
-        }
-        return newMarkers
-      }
-    }
-    return markers
-  },
-})
-
-function blameGutterExt(): Extension {
-  return [
-    blameMarkerField,
-    gutter({
-      class: 'cm-blame-gutter',
-      lineMarkerChange(update: ViewUpdate) {
-        return update.startState.field(blameMarkerField, false) !== update.state.field(blameMarkerField, false)
-      },
-      lineMarker(view, block) {
-        const line = view.state.doc.lineAt(block.from)
-        const field = view.state.field(blameMarkerField, false)
-        return field?.[line.number - 1] ?? BLAME_EMPTY
-      },
-      initialSpacer: () => BLAME_EMPTY,
-    }),
-  ]
-}
-
-export function pushOwnership(view: EditorView, ownership: ReadonlyMap<number, LineOwnership>): void {
-  view.dispatch({ effects: [setOwnership.of(ownership)] })
-}
-
-// ── Keeper trace gutter ───────────────────────────────────────────
-// File-scoped line trace dots for the `keeper-trace` layer.
-
-const TRACE_DOT_CAP = 3
-
-export interface EditorKeeperTraceLineEvent {
-  readonly id?: string
-  readonly source: KeeperTraceSource
-  readonly keeperName: string
-  readonly count: number
-  readonly tsMs: number
-  readonly filePath?: string
-  readonly line?: number
-  readonly eventId?: string
-  readonly threadId?: string
-  readonly surface?: string
-  readonly goalId?: string
-  readonly taskId?: string
-  readonly boardPostId?: string
-  readonly commentId?: string
-  readonly prId?: string
-  readonly gitRef?: string
-  readonly logId?: string
-  readonly sessionId?: string
-  readonly operationId?: string
-  readonly workerRunId?: string
-  readonly decisionChoice?: string | null
-  readonly decisionReason?: string | null
-}
-
-export interface EditorKeeperTraceLine {
-  readonly line: number
-  readonly events: ReadonlyArray<EditorKeeperTraceLineEvent>
-}
-
-export interface KeeperTraceLineGutterOptions {
-  readonly getTraceLines?: () => ReadonlyArray<EditorKeeperTraceLine>
-  readonly onTraceLineSelect?: (event: EditorKeeperTraceLineEvent, line: number) => void
-}
-
-const setKeeperTraceLines = StateEffect.define<ReadonlyArray<EditorKeeperTraceLine>>()
-
-class TraceLineMarker extends GutterMarker {
-  constructor(
-    private readonly line: number,
-    private readonly events: ReadonlyArray<EditorKeeperTraceLineEvent>,
-  ) {
-    super()
-  }
-
-  toDOM(): HTMLElement {
-    if (this.events.length === 0) {
-      const el = document.createElement('span')
-      el.className = 'cm-trace-stack'
-      el.setAttribute('aria-hidden', 'true')
-      return el
-    }
-    const el = document.createElement('button')
-    el.type = 'button'
-    el.className = 'cm-trace-stack'
-    el.dataset.line = String(this.line)
-    el.setAttribute('aria-label', traceLineAriaLabel(this.line, this.events))
-    el.title = traceLineTitle(this.line, this.events)
-    const visible = this.events.slice(0, TRACE_DOT_CAP)
-    for (const event of visible) {
-      const dot = document.createElement('span')
-      dot.className = 'cm-trace-dot'
-      dot.setAttribute('role', 'img')
-      dot.setAttribute('aria-label', traceEventLabel(event))
-      dot.setAttribute('data-source', event.source)
-      dot.style.setProperty('--cm-trace-color', traceSourceColor(event.source))
-      el.append(dot)
-    }
-    const overflow = this.events.length - visible.length
-    if (overflow > 0) {
-      const more = document.createElement('span')
-      more.className = 'cm-trace-overflow'
-      more.textContent = `+${overflow}`
-      more.setAttribute('aria-label', `${overflow} more trace events`)
-      el.append(more)
-    }
-    return el
-  }
-
-  eq(other: TraceLineMarker): boolean {
-    return this.line === other.line && traceLineKey(this.events) === traceLineKey(other.events)
-  }
-}
-
-const TRACE_SPACER = new TraceLineMarker(0, [])
-
-const keeperTraceLineField = StateField.define<ReadonlyMap<number, TraceLineMarker>>({
-  create() {
-    return new Map()
-  },
-  update(markers, tr) {
-    for (const effect of tr.effects) {
-      if (!effect.is(setKeeperTraceLines)) continue
-      const next = new Map<number, TraceLineMarker>()
-      for (const traceLine of effect.value) {
-        if (traceLine.line < 1 || traceLine.line > tr.state.doc.lines || traceLine.events.length === 0) continue
-        next.set(traceLine.line, new TraceLineMarker(traceLine.line, traceLine.events))
-      }
-      return next
-    }
-    return markers
-  },
-})
-
-class TraceLineChip extends WidgetType {
-  constructor(private readonly traceLine: EditorKeeperTraceLine) {
-    super()
-  }
-
-  toDOM(): HTMLElement {
-    const el = document.createElement('span')
-    const first = this.traceLine.events[0]
-    el.className = 'cm-masc-trace-chip'
-    el.textContent = traceLineChipText(this.traceLine)
-    el.title = traceLineChipTitle(this.traceLine)
-    el.setAttribute('aria-label', traceLineChipAriaLabel(this.traceLine))
-    if (first) el.style.setProperty('--cm-trace-chip-color', traceSourceColor(first.source))
-    return el
-  }
-
-  eq(other: TraceLineChip): boolean {
-    return this.traceLine.line === other.traceLine.line
-      && traceLineKey(this.traceLine.events) === traceLineKey(other.traceLine.events)
-  }
-}
-
-const keeperTraceLineChipField = StateField.define<DecorationSet>({
-  create() {
-    return Decoration.none
-  },
-  update(value, tr) {
-    const mapped = value.map(tr.changes)
-    for (const effect of tr.effects) {
-      if (!effect.is(setKeeperTraceLines)) continue
-      return buildTraceLineChipDecorations(tr.state.doc, effect.value)
-    }
-    return mapped
-  },
-  provide: field => EditorView.decorations.from(field),
-})
-
-export function keeperTraceLineGutterExt(options: KeeperTraceLineGutterOptions = {}): Extension {
-  const canSelect = options.getTraceLines && options.onTraceLineSelect
-  return [
-    keeperTraceLineField,
-    gutter({
-      class: 'cm-trace-gutter',
-      lineMarkerChange(update: ViewUpdate) {
-        return update.startState.field(keeperTraceLineField, false) !== update.state.field(keeperTraceLineField, false)
-      },
-      lineMarker(view, block) {
-        const line = view.state.doc.lineAt(block.from)
-        const field = view.state.field(keeperTraceLineField, false)
-        return field?.get(line.number) ?? null
-      },
-      ...(canSelect
-        ? {
-            domEventHandlers: {
-              click: (view, block, event) => selectTraceLineFromGutterClick(
-                view.state.doc.lineAt(block.from).number,
-                event,
-                options.getTraceLines!,
-                options.onTraceLineSelect!,
-              ),
-            },
-          }
-        : {}),
-      initialSpacer: () => TRACE_SPACER,
-    }),
-  ]
-}
-
-export function keeperTraceLineChipExt(): Extension {
-  return keeperTraceLineChipField
-}
-
-function selectTraceLineFromGutterClick(
-  blockLine: number,
-  event: Event,
-  getTraceLines: () => ReadonlyArray<EditorKeeperTraceLine>,
-  onTraceLineSelect: (event: EditorKeeperTraceLineEvent, line: number) => void,
-): boolean {
-  if (!(event instanceof MouseEvent) || event.button !== 0) return false
-  const target = event.target
-  if (!(target instanceof Element)) return false
-  const stack = target.closest('.cm-trace-stack')
-  if (!(stack instanceof HTMLElement) || stack.getAttribute('aria-hidden') === 'true') return false
-  const lineFromMarker = Number(stack.dataset.line)
-  const line = isPositiveSafeInteger(lineFromMarker) ? lineFromMarker : blockLine
-  const traceLine = getTraceLines().find(candidate => candidate.line === line)
-  const topEvent = traceLine?.events[0]
-  if (!topEvent) return false
-  event.stopPropagation()
-  onTraceLineSelect(topEvent, line)
-  return true
-}
-
-export function pushKeeperTraceLines(
-  view: EditorView,
-  traceLines: ReadonlyArray<EditorKeeperTraceLine>,
-): void {
-  view.dispatch({ effects: [setKeeperTraceLines.of(traceLines)] })
-}
-
-export function keeperTraceLinesForFile(
-  filePath: string,
-  events: ReadonlyArray<KeeperTraceEvent>,
-): ReadonlyArray<EditorKeeperTraceLine> {
-  const normalizedFilePath = filePath.trim()
-  if (!normalizedFilePath) return []
-
-  const byLine = new Map<number, EditorKeeperTraceLineEvent[]>()
-  for (const event of events) {
-    const filePath = traceEventFilePath(event)
-    const line = traceEventLine(event)
-    if (filePath !== normalizedFilePath) continue
-    if (line === null) continue
-    const existing = byLine.get(line) ?? []
-    const context = traceEventContextFields(event)
-    existing.push({
-      id: event.id,
-      source: event.source,
-      keeperName: event.keeperName,
-      count: event.count,
-      tsMs: event.tsMs,
-      filePath,
-      line,
-      eventId: event.source === 'activity-event' ? event.eventId : undefined,
-      threadId: event.source === 'anchored-thread' ? event.threadId : undefined,
-      surface: traceEventSurface(event),
-      goalId: context.goalId,
-      taskId: context.taskId,
-      boardPostId: context.boardPostId,
-      commentId: context.commentId,
-      prId: context.prId,
-      gitRef: context.gitRef,
-      logId: context.logId,
-      sessionId: context.sessionId,
-      operationId: context.operationId,
-      workerRunId: context.workerRunId,
-      decisionChoice: event.source === 'decision-log' ? event.decisionChoice : undefined,
-      decisionReason: event.source === 'decision-log' ? event.decisionReason : undefined,
-    })
-    byLine.set(line, existing)
-  }
-
-  return [...byLine.entries()]
-    .sort(([left], [right]) => left - right)
-    .map(([line, eventsForLine]) => ({
-      line,
-      events: eventsForLine.sort((left, right) => right.tsMs - left.tsMs),
-    }))
-}
-
-function buildTraceLineChipDecorations(
-  doc: Text,
-  traceLines: ReadonlyArray<EditorKeeperTraceLine>,
-): DecorationSet {
-  const builder = new RangeSetBuilder<Decoration>()
-  const sorted = [...traceLines]
-    .filter(traceLine =>
-      traceLine.line >= 1
-      && traceLine.line <= doc.lines
-      && traceLine.events.length > 0,
-    )
-    .sort((left, right) => left.line - right.line)
-
-  for (const traceLine of sorted) {
-    const line = doc.line(traceLine.line)
-    builder.add(
-      line.to,
-      line.to,
-      Decoration.widget({
-        widget: new TraceLineChip(traceLine),
-        side: 3,
-      }),
-    )
-  }
-
-  return builder.finish()
-}
-
-function traceSourceColor(source: KeeperTraceSource): string {
-  switch (source) {
-    case 'anchored-thread':
-      return 'var(--color-status-info)'
-    case 'activity-event':
-      return 'var(--color-status-info)'
-    case 'runtime-hop':
-      return 'var(--color-accent-fg)'
-    case 'bdi-snapshot':
-      return 'var(--color-status-ok)'
-    case 'decision-log':
-      return 'var(--color-status-warn)'
-  }
-}
-
-function traceSourceLabel(source: KeeperTraceSource): string {
-  switch (source) {
-    case 'anchored-thread':
-      return 'thread'
-    case 'activity-event':
-      return 'activity'
-    case 'runtime-hop':
-      return 'runtime'
-    case 'bdi-snapshot':
-      return 'BDI'
-    case 'decision-log':
-      return 'decision'
-  }
-}
-
-function traceEventFilePath(event: KeeperTraceEvent): string | null {
-  if (event.source === 'anchored-thread') return event.filePath ?? null
-  if (event.source === 'activity-event') return event.filePath
-  if (
-    event.source === 'runtime-hop'
-    || event.source === 'bdi-snapshot'
-    || event.source === 'decision-log'
-  ) return event.filePath ?? null
-  return null
-}
-
-function traceEventLine(event: KeeperTraceEvent): number | null {
-  if (event.source === 'anchored-thread') {
-    return isPositiveSafeInteger(event.line) ? event.line : null
-  }
-  if (event.source === 'activity-event') {
-    return isPositiveSafeInteger(event.line) ? event.line : null
-  }
-  if (
-    event.source === 'runtime-hop'
-    || event.source === 'bdi-snapshot'
-    || event.source === 'decision-log'
-  ) {
-    const line = event.line
-    return isPositiveSafeInteger(line)
-      ? line
-      : null
-  }
-  return null
-}
-
-function traceEventSurface(event: KeeperTraceEvent): string | undefined {
-  return event.source === 'activity-event' ? event.surface : undefined
-}
-
-function traceEventContextFields(event: KeeperTraceEvent): KeeperTraceContextFields {
-  if (
-    event.source === 'activity-event'
-    || event.source === 'runtime-hop'
-    || event.source === 'bdi-snapshot'
-    || event.source === 'decision-log'
-  ) return event
-  return {}
-}
-
-function traceEventLabel(event: EditorKeeperTraceLineEvent): string {
-  const count = event.count > 1 ? ` x${event.count}` : ''
-  const rawSurface = event.surface?.trim()
-  const surface = event.source === 'activity-event'
-    && rawSurface
-    && rawSurface.toLowerCase() !== 'activity'
-    ? ` ${rawSurface}`
-    : ''
-  return `${traceSourceLabel(event.source)}${surface} ${event.keeperName}${count}`
-}
-
-function traceLineChipText(traceLine: EditorKeeperTraceLine): string {
-  const first = traceLine.events[0]
-  if (!first) return 'Trace'
-  const parts = [
-    'Trace',
-    traceLineChipSurface(first),
-    ...traceEventContextParts(first),
-    `keeper ${first.keeperName}`,
-    traceLine.events.length > 1 ? `+${traceLine.events.length - 1}` : null,
-  ].filter((part): part is string => part !== null)
-  return parts.join(' · ')
-}
-
-function traceLineChipTitle(traceLine: EditorKeeperTraceLine): string {
-  return traceLine.events
-    .map(event => [
-      traceLineChipSurface(event),
-      traceLineChipLabel(event),
-      ...traceEventContextParts(event),
-      `keeper ${event.keeperName}`,
-    ].filter((part): part is string => part !== null).join(' · '))
-    .join('\n')
-}
-
-function traceLineChipAriaLabel(traceLine: EditorKeeperTraceLine): string {
-  return `Line ${traceLine.line} keeper trace context: ${traceLineChipText(traceLine)}`
-}
-
-function traceLineChipSurface(event: EditorKeeperTraceLineEvent): string {
-  if (event.source === 'activity-event') return event.surface?.trim() || 'Activity'
-  if (event.source === 'anchored-thread') return 'Thread'
-  if (event.source === 'runtime-hop') return 'Runtime'
-  if (event.source === 'bdi-snapshot') return 'BDI'
-  return 'Decision'
-}
-
-function traceLineChipLabel(event: EditorKeeperTraceLineEvent): string {
-  if (event.source === 'activity-event') {
-    const surface = traceLineChipSurface(event)
-    return event.eventId ? `${surface} activity ${event.eventId}` : surface
-  }
-  if (event.source === 'anchored-thread') {
-    return event.threadId ? `thread ${event.threadId}` : 'thread'
-  }
-  if (event.source === 'decision-log') {
-    const choice = event.decisionChoice?.trim()
-    const reason = event.decisionReason?.trim()
-    if (choice && reason) return `${choice}: ${reason}`
-    if (choice) return choice
-    if (reason) return reason
-  }
-  return traceLineChipSurface(event)
-}
-
-function traceEventContextParts(event: EditorKeeperTraceLineEvent): ReadonlyArray<string> {
-  return [
-    event.eventId ? `event ${event.eventId}` : null,
-    event.threadId ? `thread ${event.threadId}` : null,
-    event.goalId ? `goal ${event.goalId}` : null,
-    event.taskId ? `task ${event.taskId}` : null,
-    event.boardPostId ? `board ${event.boardPostId}` : null,
-    event.commentId ? `comment ${event.commentId}` : null,
-    event.prId ? `pr #${event.prId}` : null,
-    event.gitRef ? `git ${shortGitRef(event.gitRef)}` : null,
-    event.logId ? `log ${event.logId}` : null,
-    event.sessionId ? `session ${event.sessionId}` : null,
-    event.operationId ? `op ${event.operationId}` : null,
-    event.workerRunId ? `run ${event.workerRunId}` : null,
-  ].filter((part): part is string => part !== null)
-}
-
-function shortGitRef(ref: string): string {
-  return ref.replace(/^refs\/heads\//, '').replace(/^refs\/tags\//, '')
-}
-
-function traceLineAriaLabel(line: number, events: ReadonlyArray<EditorKeeperTraceLineEvent>): string {
-  return `Line ${line} keeper trace: ${events.map(traceEventLabel).join(', ')}`
-}
-
-function traceLineTitle(line: number, events: ReadonlyArray<EditorKeeperTraceLineEvent>): string {
-  return [`L${line}`, ...events.map(traceEventLabel)].join('\n')
-}
-
-function traceLineKey(events: ReadonlyArray<EditorKeeperTraceLineEvent>): string {
-  return events
-    .map(event => [
-      event.id ?? '',
-      event.source,
-      event.surface ?? '',
-      event.keeperName,
-      event.count,
-      event.tsMs,
-      event.goalId ?? '',
-      event.taskId ?? '',
-      event.boardPostId ?? '',
-      event.commentId ?? '',
-      event.prId ?? '',
-      event.gitRef ?? '',
-      event.logId ?? '',
-      event.sessionId ?? '',
-      event.operationId ?? '',
-      event.workerRunId ?? '',
-    ].join(':'))
-    .join('|')
-}
+// ── Keeper line selection handler ─────────────────────────────────
 
 export function keeperLineSelectExt(
   getOwnership: () => ReadonlyMap<number, LineOwnership>,
@@ -1049,381 +284,8 @@ function annotationLineKey(annotations: ReadonlyArray<EditorAnnotationLine>): st
   ).join('|')
 }
 
-// ── Language support (dynamic import) ─────────────────────────────
-
-type LanguageModule = () => Promise<{ extension: Extension }>
-
-interface LanguageEntry {
-  readonly id: string
-  readonly load: LanguageModule
-}
-
-interface OcamlStreamState {
-  commentDepth: number
-  stringQuote: '"' | null
-}
-
-const OCAML_KEYWORDS = new Set([
-  'and',
-  'as',
-  'assert',
-  'begin',
-  'class',
-  'constraint',
-  'do',
-  'done',
-  'downto',
-  'else',
-  'end',
-  'exception',
-  'external',
-  'for',
-  'fun',
-  'function',
-  'functor',
-  'if',
-  'in',
-  'include',
-  'inherit',
-  'initializer',
-  'lazy',
-  'let',
-  'match',
-  'method',
-  'module',
-  'mutable',
-  'new',
-  'nonrec',
-  'object',
-  'of',
-  'open',
-  'private',
-  'rec',
-  'sig',
-  'struct',
-  'then',
-  'to',
-  'try',
-  'type',
-  'val',
-  'virtual',
-  'when',
-  'while',
-  'with',
-])
-
-const OCAML_ATOMS = new Set(['false', 'true', 'None', 'Some', 'Ok', 'Error'])
-
-const OCAML_BUILTINS = new Set([
-  'bool',
-  'char',
-  'exn',
-  'float',
-  'int',
-  'list',
-  'option',
-  'result',
-  'string',
-  'unit',
-])
-
-const ocamlStreamParser: StreamParser<OcamlStreamState> = {
-  name: 'ocaml',
-  startState: () => ({ commentDepth: 0, stringQuote: null }),
-  copyState: state => ({ ...state }),
-  languageData: {
-    name: 'ocaml',
-    commentTokens: { block: { open: '(*', close: '*)' } },
-  },
-  token(stream, state) {
-    if (state.commentDepth > 0) return readOcamlComment(stream, state)
-    if (state.stringQuote !== null) return readOcamlString(stream, state)
-    if (stream.eatSpace()) return null
-
-    if (stream.match('(*')) {
-      state.commentDepth = 1
-      return readOcamlComment(stream, state)
-    }
-
-    const ch = stream.next()
-    if (ch === undefined) return null
-
-    if (ch === '"') {
-      state.stringQuote = '"'
-      return readOcamlString(stream, state)
-    }
-
-    if (ch === '\'') {
-      if (stream.match(/^\\?.'/)) return 'string'
-      stream.eatWhile(/[A-Za-z0-9_']/)
-      return 'typeName'
-    }
-
-    if (/[0-9]/.test(ch)) {
-      stream.eatWhile(/[A-Za-z0-9_'.]/)
-      return 'number'
-    }
-
-    if (/[A-Z]/.test(ch)) {
-      stream.eatWhile(/[A-Za-z0-9_']/)
-      const ident = stream.current()
-      return OCAML_ATOMS.has(ident) ? 'atom' : 'typeName'
-    }
-
-    if (/[a-z_]/.test(ch)) {
-      stream.eatWhile(/[A-Za-z0-9_']/)
-      const ident = stream.current()
-      if (OCAML_KEYWORDS.has(ident)) return 'keyword'
-      if (OCAML_ATOMS.has(ident)) return 'atom'
-      if (OCAML_BUILTINS.has(ident)) return 'standard variableName'
-      return 'variableName'
-    }
-
-    if (/[-+*/=<>@^|&$%!?~:.;,#]/.test(ch)) {
-      stream.eatWhile(/[-+*/=<>@^|&$%!?~:.;,#]/)
-      return 'operator'
-    }
-
-    return null
-  },
-}
-
-const ocamlLanguage = StreamLanguage.define(ocamlStreamParser)
-
-// ── TOML StreamLanguage ──────────────────────────────
-
-interface TomlStreamState {
-  readonly inSection: boolean
-}
-
-const tomlStreamParser: StreamParser<TomlStreamState> = {
-  name: 'toml',
-  startState: () => ({ inSection: false }),
-  copyState: state => ({ ...state }),
-  languageData: {
-    name: 'toml',
-    commentTokens: { line: '#' },
-  },
-  token(stream, state) {
-    if (stream.eatSpace()) return null
-
-    if (stream.sol() && stream.peek() === '#') {
-      stream.skipToEnd()
-      return 'comment'
-    }
-
-    if (stream.sol() && stream.peek() === '[') {
-      stream.skipToEnd()
-      return 'heading'
-    }
-
-    if (stream.peek() === '=') {
-      stream.next()
-      return 'operator'
-    }
-
-    if (stream.peek() === '"' || stream.peek() === '\'') {
-      const quote = stream.next()!
-      let escaped = false
-      while (!stream.eol()) {
-        const ch = stream.next()!
-        if (escaped) { escaped = false; continue }
-        if (ch === '\\') { escaped = true; continue }
-        if (ch === quote) break
-      }
-      return 'string'
-    }
-
-    if (/[0-9\-]/.test(stream.peek() ?? '') && stream.match(/^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?/)) {
-      return 'number'
-    }
-
-    if (stream.match(/^true\b/) || stream.match(/^false\b/)) {
-      return 'atom'
-    }
-
-    if (stream.match(/^\d{4}-\d{2}-\d{2}/)) {
-      stream.eatWhile(/[T :\d.Z]/)
-      return 'number'
-    }
-
-    if (/[A-Za-z_]/.test(stream.peek() ?? '')) {
-      stream.eatWhile(/[A-Za-z0-9_.\-]/)
-      return state.inSection ? 'propertyName' : 'variableName'
-    }
-
-    stream.next()
-    return null
-  },
-}
-
-const tomlLanguage = StreamLanguage.define(tomlStreamParser)
-
-// ── YAML StreamLanguage ──────────────────────────────
-
-interface YamlStreamState {
-  inKey: boolean
-}
-
-const YAML_KEYWORDS = new Set(['true', 'false', 'null', '~', 'yes', 'no', 'on', 'off'])
-
-const yamlStreamParser: StreamParser<YamlStreamState> = {
-  name: 'yaml',
-  startState: () => ({ inKey: true }),
-  copyState: state => ({ ...state }),
-  languageData: {
-    name: 'yaml',
-    commentTokens: { line: '#' },
-  },
-  token(stream, state) {
-    if (stream.eatSpace()) return null
-
-    if (stream.peek() === '#') {
-      stream.skipToEnd()
-      return 'comment'
-    }
-
-    if (stream.sol() && (stream.match(/^---/) || stream.match(/^\.\.\./))) {
-      stream.eatSpace()
-      return 'operator'
-    }
-
-    if (stream.sol() && stream.peek() === '-') {
-      stream.next()
-      if (stream.eatSpace()) return 'operator'
-    }
-
-    if (stream.peek() === '"' || stream.peek() === '\'') {
-      const quote = stream.next()!
-      let escaped = false
-      while (!stream.eol()) {
-        const ch = stream.next()!
-        if (escaped) { escaped = false; continue }
-        if (ch === '\\') { escaped = true; continue }
-        if (ch === quote) break
-      }
-      state.inKey = false
-      return 'string'
-    }
-
-    if (stream.peek() === ':') {
-      stream.next()
-      if (stream.eatSpace()) {
-        state.inKey = true
-        return 'operator'
-      }
-      return null
-    }
-
-    if (/[0-9\-]/.test(stream.peek() ?? '') && stream.match(/^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?/)) {
-      state.inKey = false
-      return 'number'
-    }
-
-    if (/[A-Za-z_]/.test(stream.peek() ?? '')) {
-      stream.eatWhile(/[A-Za-z0-9_.\-]/)
-      const word = stream.current()
-      if (YAML_KEYWORDS.has(word)) {
-        state.inKey = false
-        return 'atom'
-      }
-      if (state.inKey) return 'propertyName'
-      state.inKey = false
-      return 'string'
-    }
-
-    if (stream.peek() === '[' || stream.peek() === ']' || stream.peek() === '{' || stream.peek() === '}' || stream.peek() === ',') {
-      stream.next()
-      return 'operator'
-    }
-
-    stream.next()
-    state.inKey = false
-    return null
-  },
-}
-
-const yamlLanguage = StreamLanguage.define(yamlStreamParser)
-
-function readOcamlComment(stream: StringStream, state: OcamlStreamState): string {
-  while (!stream.eol()) {
-    if (stream.match('(*')) {
-      state.commentDepth += 1
-      continue
-    }
-    if (stream.match('*)')) {
-      state.commentDepth -= 1
-      if (state.commentDepth <= 0) {
-        state.commentDepth = 0
-        break
-      }
-      continue
-    }
-    stream.next()
-  }
-  return 'comment'
-}
-
-function readOcamlString(stream: StringStream, state: OcamlStreamState): string {
-  let escaped = false
-  while (!stream.eol()) {
-    const ch = stream.next()
-    if (escaped) {
-      escaped = false
-    } else if (ch === '\\') {
-      escaped = true
-    } else if (ch === state.stringQuote) {
-      state.stringQuote = null
-      break
-    }
-  }
-  return 'string'
-}
-
-const LANGUAGE_MAP: Readonly<Record<string, LanguageEntry>> = {
-  '.ts': { id: 'typescript', load: () => import('@codemirror/lang-javascript').then(m => m.javascript({ typescript: true })) },
-  '.tsx': { id: 'typescript', load: () => import('@codemirror/lang-javascript').then(m => m.javascript({ typescript: true, jsx: true })) },
-  '.js': { id: 'javascript', load: () => import('@codemirror/lang-javascript').then(m => m.javascript()) },
-  '.jsx': { id: 'javascript', load: () => import('@codemirror/lang-javascript').then(m => m.javascript({ jsx: true })) },
-  '.py': { id: 'python', load: () => import('@codemirror/lang-python').then(m => m.python()) },
-  '.html': { id: 'html', load: () => import('@codemirror/lang-html').then(m => m.html()) },
-  '.css': { id: 'css', load: () => import('@codemirror/lang-css').then(m => m.css()) },
-  '.json': { id: 'json', load: () => import('@codemirror/lang-json').then(m => m.json()) },
-  '.md': { id: 'markdown', load: () => import('@codemirror/lang-markdown').then(m => m.markdown()) },
-  '.ocaml': { id: 'ocaml', load: () => Promise.resolve(ocamlLanguage) },
-  '.ml': { id: 'ocaml', load: () => Promise.resolve(ocamlLanguage) },
-  '.mli': { id: 'ocaml', load: () => Promise.resolve(ocamlLanguage) },
-  '.rs': { id: 'rust', load: () => import('@codemirror/lang-rust').then(m => m.rust()) },
-  '.go': { id: 'go', load: () => import('@codemirror/lang-go').then(m => m.go()) },
-  '.toml': { id: 'toml', load: () => Promise.resolve(tomlLanguage) },
-  '.yaml': { id: 'yaml', load: () => Promise.resolve(yamlLanguage) },
-  '.yml': { id: 'yaml', load: () => Promise.resolve(yamlLanguage) },
-}
-
-export function languageIdForFilePath(filePath: string): string | null {
-  const ext = filePath.slice(filePath.lastIndexOf('.'))
-  return LANGUAGE_MAP[ext]?.id ?? null
-}
-
-export async function languageExt(filePath: string): Promise<Extension> {
-  const ext = filePath.slice(filePath.lastIndexOf('.'))
-  const loader = LANGUAGE_MAP[ext]?.load
-  if (!loader) return []
-  try {
-    return await loader()
-  } catch {
-    return []
-  }
-}
-
 // ── Line number gutter ────────────────────────────────────────────
 
 export function lineNumberExt(): Extension {
   return lineNumbers()
-}
-
-// ── Blame mode extensions bundle ──────────────────────────────────
-
-export function blameExtensions(): Extension[] {
-  return [blameGutterExt()]
 }
