+++ dashboard/src/components/ide/ide-editor-theme.ts
@@ -1,0 +1,200 @@
+import { EditorView } from '@codemirror/view'
+import { Extension } from '@codemirror/state'
+import { defaultHighlightStyle, syntaxHighlighting } from '@codemirror/language'
+
+export function themeExt(): Extension {
+  return EditorView.theme({
+    '&': {
+      background: 'var(--color-bg-page)',
+      color: 'var(--color-fg-secondary)',
+      fontFamily: 'var(--font-mono)',
+      fontSize: 'var(--fs-13)',
+      lineHeight: '1.6',
+      height: '100%',
+    },
+    '.cm-scroller': {
+      overflow: 'auto',
+      minHeight: '0',
+    },
+    '.cm-content': {
+      caretColor: 'transparent',
+      padding: 'var(--sp-2) 0',
+      minHeight: '100%',
+    },
+    '.cm-line': {
+      padding: '0 var(--sp-3)',
+    },
+    '.cm-gutters': {
+      background: 'var(--color-bg-page)',
+      color: 'var(--color-fg-disabled)',
+      border: 'none',
+      fontSize: 'var(--fs-11)',
+      minWidth: '44px',
+      position: 'sticky',
+      left: '0',
+      zIndex: '2',
+    },
+    '.cm-gutterElement': {
+      textAlign: 'right',
+      paddingRight: 'var(--sp-2)',
+    },
+    '.cm-lineNumbers': {
+      borderRight: '1px solid var(--color-border-default)',
+    },
+    '.cm-blame-gutter': {
+      minWidth: '78px',
+      borderRight: '1px solid var(--color-border-default)',
+      background: 'var(--color-bg-surface)',
+    },
+    '.cm-blame-gutter .cm-gutterElement': {
+      padding: '0 var(--sp-2)',
+      textAlign: 'left',
+    },
+    '.cm-blame-marker': {
+      display: 'inline-grid',
+      gridTemplateColumns: '18px minmax(0, 1fr)',
+      alignItems: 'center',
+      gap: 'var(--sp-1)',
+      width: '68px',
+      minWidth: '0',
+      color: 'var(--cm-blame-color)',
+      fontSize: 'var(--fs-10)',
+      lineHeight: '1.4',
+    },
+    '.cm-blame-sigil': {
+      display: 'inline-flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      width: '16px',
+      height: '14px',
+      borderRadius: 'var(--r-0)',
+      color: 'var(--color-bg-page)',
+      background: 'var(--cm-blame-color)',
+      fontSize: 'var(--fs-9)',
+      fontWeight: '700',
+      letterSpacing: '0',
+    },
+    '.cm-blame-name': {
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      whiteSpace: 'nowrap',
+    },
+    '.cm-trace-gutter': {
+      minWidth: '34px',
+      borderRight: '1px solid var(--color-border-default)',
+      background: 'var(--color-bg-page)',
+    },
+    '.cm-trace-gutter .cm-gutterElement': {
+      padding: '0 var(--sp-1)',
+      textAlign: 'left',
+    },
+    '.cm-trace-stack': {
+      display: 'inline-flex',
+      alignItems: 'center',
+      justifyContent: 'flex-start',
+      gap: '2px',
+      minWidth: '28px',
+      height: '100%',
+      padding: '0',
+      border: '0',
+      background: 'transparent',
+      color: 'inherit',
+      cursor: 'pointer',
+    },
+    '.cm-trace-stack:focus-visible': {
+      outline: '1px solid var(--color-accent-fg)',
+      outlineOffset: '1px',
+    },
+    '.cm-trace-dot': {
+      display: 'inline-block',
+      width: '6px',
+      height: '6px',
+      borderRadius: '999px',
+      background: 'var(--cm-trace-color)',
+      boxShadow: '0 0 0 1px var(--color-bg-page)',
+    },
+    '.cm-trace-overflow': {
+      color: 'var(--color-fg-muted)',
+      fontSize: 'var(--fs-9)',
+      lineHeight: '1',
+    },
+    '&.cm-focused': {
+      outline: 'none',
+    },
+    '.cm-cursor': {
+      display: 'none',
+    },
+    '.cm-selectionBackground': {
+      background: 'transparent !important',
+    },
+    '.cm-line.cm-masc-context-focus': {
+      background: 'color-mix(in srgb, var(--color-accent-fg) 12%, transparent)',
+      boxShadow: 'inset 2px 0 0 var(--color-accent-fg)',
+    },
+    '.cm-masc-context-focus-chip': {
+      display: 'inline-flex',
+      alignItems: 'center',
+      maxWidth: '36ch',
+      marginLeft: 'var(--sp-2)',
+      padding: '0 var(--sp-2)',
+      border: '1px solid var(--color-border-default)',
+      borderRadius: 'var(--r-0)',
+      background: 'var(--color-bg-elevated)',
+      color: 'var(--color-accent-fg)',
+      fontFamily: 'var(--font-mono)',
+      fontSize: 'var(--fs-10)',
+      lineHeight: '1.4',
+      verticalAlign: 'baseline',
+      whiteSpace: 'nowrap',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      pointerEvents: 'none',
+      userSelect: 'none',
+    },
+    '.cm-masc-annotation-chip': {
+      display: 'inline-flex',
+      alignItems: 'center',
+      maxWidth: '34ch',
+      marginLeft: 'var(--sp-2)',
+      padding: '0 var(--sp-2)',
+      border: '1px solid var(--color-border-default)',
+      borderRadius: 'var(--r-0)',
+      background: 'var(--color-bg-surface)',
+      color: 'var(--color-fg-muted)',
+      fontFamily: 'var(--font-mono)',
+      fontSize: 'var(--fs-10)',
+      lineHeight: '1.4',
+      verticalAlign: 'baseline',
+      whiteSpace: 'nowrap',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      pointerEvents: 'none',
+      userSelect: 'none',
+    },
+    '.cm-masc-trace-chip': {
+      display: 'inline-flex',
+      alignItems: 'center',
+      maxWidth: '46ch',
+      marginLeft: 'var(--sp-2)',
+      padding: '0 var(--sp-2)',
+      border: '1px solid var(--color-border-default)',
+      borderRadius: 'var(--r-0)',
+      background: 'var(--color-bg-elevated)',
+      boxShadow: 'inset 2px 0 0 var(--cm-trace-chip-color)',
+      color: 'var(--color-fg-secondary)',
+      fontFamily: 'var(--font-mono)',
+      fontSize: 'var(--fs-10)',
+      lineHeight: '1.4',
+      verticalAlign: 'baseline',
+      whiteSpace: 'nowrap',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      pointerEvents: 'none',
+      userSelect: 'none',
+    },
+  })
+}
+
+export function syntaxHighlightExt(): Extension {
+  return syntaxHighlighting(defaultHighlightStyle, { fallback: true })
+}
